requirements.txt
monai[nibabel, skimage, itk, tqdm, psutil, ignite]==0.8.0
monailabel==0.3.0
ninja==1.10.2.3
GeodisTK==0.1.7
FastGeodis==1.0.2
torchhaarfeatures==0.0.2
torchgaussianfilter==0.0.2
numpymaxflow==0.0.5